Split the CLI into multiple files
Tidy the CLI tool by splitting it into subfiles.
