+++ Makefile
@@ -1,3 +1,3 @@
 default: gopenqa
-gopenqa: cmd/gopenqa/gopenqa.go *.go
-	go build -o gopenqa cmd/gopenqa/gopenqa.go
+gopenqa: cmd/gopenqa/*.go *.go
+	go build -o gopenqa cmd/gopenqa/*.go
